Fully replaced emit with iced_x86. Added Colored ASM output. Added support for GCC linking.
