+++ ReadMe.md
@@ -4,13 +4,13 @@ This project is a simple C compiler written in Rust, following Sandler's "Writin
 Some design decisions are informed by the book but the implementation is my own.
 
 So far chapter 4 (inc. bitwise extra credit) is implemented, which includes a lexer, parser, and code generator for very basic C code.
+This compiler is a fully standalone executable; it does not rely on any external programs for assembling or linking (on linux).
 All provided tests pass.
 
 ## Requirements
 
 - Rust (latest stable)
- - GCC (for the default assembly-based backend)
- - [wild](https://github.com/davidlattimore/wild) (via `libwild` crate) for the `--iced` option
+- GCC (Optional) for linking instead of `libwild`
 
 ## Building
 
@@ -18,13 +18,6 @@ Clone the repository and build with Cargo:
 
 ```sh
 cargo build --release
-```
-
-If Cargo fails with an error about `symbolic-common` not matching any
-available versions, update the crates index and lock file:
-
-```sh
-cargo update
 ```
 
 ### Running Tests
@@ -47,14 +40,14 @@ Usage: ncc [OPTIONS] `FILENAME`
 | `--codegen`               | Run lexer, parser, and code generator |
 | `--tacky`                 | Emit TACKY IR                         |
 | `--run`                   | Run Compiled Program and print result |
-| `--iced`                  | Use iced-x86 backend with wild linker |
+| `--gcc`                   | Use GCC for linking (instead of wild) |
 | `-S`                      | Emits assembly                        |
 | `-o`, `--output <OUTPUT>` | Override output file location         |
 | `-h`, `--help`            | Print help                            |
 | `-V`, `--version`         | Print version                         |
 
 
-Note: `--lex`, `--parse`, `--code-gen`, `--tacky`, `--run`, `--iced` are mutually exclusive options.
+Note: `--lex`, `--parse`, `--codegen`, `--tacky`, `--run` are mutually exclusive options.
 
 ## Contributing
 As a reminder to myself.
