+++ src/emit.rs
@@ -1,145 +1,263 @@
-use crate::codegen;
+// Simple emitter using iced-x86 to generate machine code at runtime
+use crate::codegen::{self, BinaryOp, CondCode, Instruction, Operand, Reg, UnaryOp};
+use iced_x86::BlockEncoderOptions;
+use iced_x86::IcedError;
+use iced_x86::code_asm::*;
+use object::write::{
+    Object, StandardSection, StandardSegment, Symbol, SymbolFlags, SymbolKind, SymbolScope, SymbolSection,
+};
+use object::{Architecture, BinaryFormat, Endianness, SectionKind};
+use std::collections::HashMap;
 
-enum RegWidth {
-    Byte,
-    DWord,
-}
-fn emit_reg(reg: &codegen::Reg, reg_width: &RegWidth) -> String {
-    match reg_width {
-        RegWidth::Byte => match reg {
-            codegen::Reg::AX => "al".to_string(),
-            codegen::Reg::R10 => "r10b".to_string(),
-            codegen::Reg::R11 => "r11b".to_string(),
-            codegen::Reg::DX => "dl".to_string(),
-            codegen::Reg::CX => "cl".to_string(),
-        },
-        RegWidth::DWord => match reg {
-            codegen::Reg::AX => "eax".to_string(),
-            codegen::Reg::R10 => "r10d".to_string(),
-            codegen::Reg::R11 => "r11d".to_string(),
-            codegen::Reg::DX => "edx".to_string(),
-            codegen::Reg::CX => "ecx".to_string(),
-        },
-    }
+pub fn get_instructions(program: &codegen::Program) -> Result<Vec<iced_x86::Instruction>, IcedError> {
+    let mut a = CodeAssembler::new(64)?;
+    emit_function(&mut a, &program.function)?;
+    Ok(Vec::from(a.instructions()))
 }
 
-fn emit_unaryop(op: &codegen::UnaryOp) -> String {
-    match op {
-        codegen::UnaryOp::Neg => "negl".to_string(),
-        codegen::UnaryOp::Not => "notl".to_string(),
-    }
-}
+pub fn emit_object(program: &codegen::Program) -> Result<Vec<u8>, Box<dyn std::error::Error>> {
+    let mut a = CodeAssembler::new(64)?;
+    let mut start_lbl = a.create_label();
+    let mut main_lbl = a.create_label();
+    a.set_label(&mut start_lbl)?;
+    a.call(main_lbl)?;
+    a.mov(gpr64::rdi, gpr64::rax)?;
+    a.mov(gpr64::rax, 60i64)?;
+    a.syscall()?;
+    a.set_label(&mut main_lbl)?;
+    emit_function(&mut a, &program.function)?;
+    let result = a.assemble_options(0, BlockEncoderOptions::RETURN_NEW_INSTRUCTION_OFFSETS)?;
+    let start_off = result.label_ip(&start_lbl)?;
+    let main_off = result.label_ip(&main_lbl)?;
+    let code = result.inner.code_buffer;
+    let start_size = main_off - start_off;
+    let main_size = code.len() as u64 - main_off;
 
-fn emit_binaryop(op: &codegen::BinaryOp) -> String {
-    match op {
-        codegen::BinaryOp::Add => "addl".to_string(),
-        codegen::BinaryOp::Sub => "subl".to_string(),
-        codegen::BinaryOp::Mult => "imull".to_string(),
-        codegen::BinaryOp::BitAnd => "andl".to_string(),
-        codegen::BinaryOp::BitOr => "orl".to_string(),
-        codegen::BinaryOp::BitXOr => "xorl".to_string(),
-        codegen::BinaryOp::BitShl => "shll".to_string(),
-        codegen::BinaryOp::BitSar => "sarl".to_string(),
-    }
-}
+    let mut obj = if cfg!(target_os = "linux") {
+        let mut obj = Object::new(BinaryFormat::Elf, Architecture::X86_64, Endianness::Little);
 
-fn emit_operand(operand: &codegen::Operand, reg_width: &RegWidth) -> String {
-    let mut output = String::new();
-    match operand {
-        codegen::Operand::Imm(value) => {
-            output.push_str(&format!("${value}"));
-        }
-        codegen::Operand::Reg(reg) => {
-            output.push_str(&format!("%{}", emit_reg(reg, reg_width)));
-        }
-        codegen::Operand::Stack(offset) => {
-            output.push_str(&format!("{offset}(%rbp)"));
-        }
-        &codegen::Operand::Pseudo(_) => unreachable!("Must be eliminated before emission"),
-    }
-    output
-}
-
-fn emit_instruction(ins: &codegen::Instruction) -> String {
-    fn emit_operand_dw(operand: &codegen::Operand) -> String {
-        emit_operand(operand, &RegWidth::DWord)
-    }
-    let mut output = String::new();
-    match ins {
-        codegen::Instruction::Mov { src, dst } => {
-            output.push_str(&format!("movl {}, {}", emit_operand_dw(src), emit_operand_dw(dst)));
-        }
-        codegen::Instruction::Ret => {
-            output.push_str("movq %rbp, %rsp\n");
-            output.push_str("\tpopq %rbp\n");
-            output.push_str("\tret");
-        }
-        codegen::Instruction::Unary { op, dst } => {
-            output.push_str(&format!("{} {}\n", emit_unaryop(op), emit_operand_dw(dst)));
-        }
-        codegen::Instruction::AllocateStack(offset) => {
-            output.push_str(&format!("subq ${}, %rsp", offset * -1));
-        }
-        codegen::Instruction::Binary { op, src, dst } => {
-            output.push_str(&format!(
-                "{} {}, {}",
-                emit_binaryop(op),
-                emit_operand_dw(src),
-                emit_operand_dw(dst)
-            ));
-        }
-        codegen::Instruction::Idiv(op) => {
-            output.push_str(&format!("idivl {} ", emit_operand_dw(op)));
-        }
-        codegen::Instruction::Cdq => {
-            output.push_str("cdq");
-        }
-        codegen::Instruction::Cmp { v1, v2 } => {
-            output.push_str(&format!("cmpl {}, {}", emit_operand_dw(v1), emit_operand_dw(v2)));
-        }
-        codegen::Instruction::Jmp(target) => {
-            output.push_str(&format!("jmp {target}"));
-        }
-        codegen::Instruction::JmpCC { code, label } => {
-            output.push_str(&format!("j{} {}", code.ins_suffix(), label));
-        }
-        codegen::Instruction::Label(label) => {
-            output.push_str(&format!("{label}:"));
-        }
-        codegen::Instruction::SetCC { code, op } => {
-            output.push_str(&format!(
-                "set{} {}",
-                code.ins_suffix(),
-                emit_operand(op, &RegWidth::Byte)
-            ));
-        }
-    }
-    output
-}
-
-fn emit_function(fun_def: &codegen::FunctionDefinition) -> String {
-    let mut output = String::new();
-    let codegen::FunctionDefinition { name, body } = fun_def;
-    let processed_name = if cfg!(target_os = "macos") {
-        format!("_{name}")
+        let note_stack = obj.add_section(
+            obj.segment_name(StandardSegment::Data).to_vec(),
+            b".note.GNU-stack".to_vec(),
+            SectionKind::Metadata,
+        );
+        obj.append_section_data(note_stack, &[], 1);
+        obj
     } else {
-        name.to_string()
+        Object::new(BinaryFormat::MachO, Architecture::X86_64, Endianness::Little)
     };
-    output.push_str(&format!("\t.global {processed_name}\n"));
-    output.push_str(&format!("{processed_name}:\n"));
-    output.push_str("\tpushq %rbp\n");
-    output.push_str("\tmovq %rsp, %rbp\n");
-    for ins in body.iter() {
-        output.push_str(&format!("\t{}\n", emit_instruction(ins)));
-    }
-    output
+
+    obj.add_file_symbol(b"ncc".to_vec());
+    let text = obj.section_id(StandardSection::Text);
+    obj.append_section_data(text, &code, 1);
+    obj.add_symbol(Symbol {
+        name: b"_start".to_vec(),
+        value: start_off,
+        size: start_size,
+        kind: SymbolKind::Text,
+        scope: SymbolScope::Linkage,
+        weak: false,
+        section: SymbolSection::Section(text),
+        flags: SymbolFlags::None,
+    });
+    obj.add_symbol(Symbol {
+        name: b"main".to_vec(),
+        value: main_off,
+        size: main_size,
+        kind: SymbolKind::Text,
+        scope: SymbolScope::Linkage,
+        weak: false,
+        section: SymbolSection::Section(text),
+        flags: SymbolFlags::None,
+    });
+    Ok(obj.write()?)
 }
 
-pub fn emit_program(program: &codegen::Program) -> String {
-    let codegen::Program { function } = program;
-    let mut output = emit_function(function);
-    if cfg!(target_os = "linux") {
-        output.push_str("\n.section .note.GNU-stack,\"\",@progbits\n");
+fn emit_function(a: &mut CodeAssembler, fun_def: &codegen::FunctionDefinition) -> Result<(), IcedError> {
+    let mut labels: HashMap<String, CodeLabel> = HashMap::new();
+    a.push(gpr64::rbp)?;
+    a.mov(gpr64::rbp, gpr64::rsp)?;
+    for ins in &fun_def.body {
+        emit_instruction(a, ins, &mut labels)?;
     }
-    output
+    Ok(())
+}
+
+fn gpr32(reg: &Reg) -> AsmRegister32 {
+    use codegen::Reg::*;
+    match reg {
+        AX => registers::gpr32::eax,
+        DX => registers::gpr32::edx,
+        R10 => registers::gpr32::r10d,
+        R11 => registers::gpr32::r11d,
+        CX => registers::gpr32::ecx,
+    }
+}
+
+fn mem_rbp(offset: i64) -> AsmMemoryOperand {
+    if offset >= 0 {
+        dword_ptr(gpr64::rbp + offset)
+    } else {
+        dword_ptr(gpr64::rbp - (-offset))
+    }
+}
+
+fn emit_instruction(
+    a: &mut CodeAssembler,
+    ins: &Instruction,
+    labels: &mut HashMap<String, CodeLabel>,
+) -> Result<(), IcedError> {
+    match ins {
+        Instruction::Mov { src, dst } => match (src, dst) {
+            (Operand::Imm(v), Operand::Reg(r)) => a.mov(gpr32(r), *v as i32)?,
+            (Operand::Reg(s), Operand::Reg(d)) => a.mov(gpr32(d), gpr32(s))?,
+            (Operand::Reg(s), Operand::Stack(off)) => a.mov(mem_rbp(*off), gpr32(s))?,
+            (Operand::Stack(off), Operand::Reg(d)) => a.mov(gpr32(d), mem_rbp(*off))?,
+            (Operand::Imm(v), Operand::Stack(off)) => a.mov(mem_rbp(*off), *v as i32)?,
+            _ => unreachable!("unsupported mov combination"),
+        },
+        Instruction::Unary { op, dst } => match op {
+            UnaryOp::Neg => match dst {
+                Operand::Reg(r) => a.neg(gpr32(r))?,
+                Operand::Stack(off) => a.neg(mem_rbp(*off))?,
+                _ => unreachable!(),
+            },
+            UnaryOp::Not => match dst {
+                Operand::Reg(r) => a.not(gpr32(r))?,
+                Operand::Stack(off) => a.not(mem_rbp(*off))?,
+                _ => unreachable!(),
+            },
+        },
+        Instruction::Binary { op, src, dst } => match op {
+            BinaryOp::Add => match (src, dst) {
+                (Operand::Reg(s), Operand::Reg(d)) => a.add(gpr32(d), gpr32(s))?,
+                (Operand::Imm(v), Operand::Reg(d)) => a.add(gpr32(d), *v as i32)?,
+                (Operand::Reg(s), Operand::Stack(off)) => a.add(mem_rbp(*off), gpr32(s))?,
+                (Operand::Imm(v), Operand::Stack(off)) => a.add(mem_rbp(*off), *v as i32)?,
+                _ => unimplemented!(),
+            },
+            BinaryOp::Sub => match (src, dst) {
+                (Operand::Reg(s), Operand::Reg(d)) => a.sub(gpr32(d), gpr32(s))?,
+                (Operand::Imm(v), Operand::Reg(d)) => a.sub(gpr32(d), *v as i32)?,
+                (Operand::Reg(s), Operand::Stack(off)) => a.sub(mem_rbp(*off), gpr32(s))?,
+                (Operand::Imm(v), Operand::Stack(off)) => a.sub(mem_rbp(*off), *v as i32)?,
+                _ => unimplemented!(),
+            },
+            BinaryOp::Mult => match (src, dst) {
+                (Operand::Reg(s), Operand::Reg(d)) => a.imul_2(gpr32(d), gpr32(s))?,
+                (Operand::Imm(v), Operand::Reg(d)) => a.imul_3(gpr32(d), gpr32(d), *v as i32)?,
+                (Operand::Reg(s), Operand::Stack(off)) => a.imul_2(gpr32(s), mem_rbp(*off))?,
+                (Operand::Stack(off), Operand::Reg(d)) => a.imul_2(gpr32(d), mem_rbp(*off))?,
+                _ => unimplemented!("Mult {:?}, {:?}", src, dst),
+            },
+            BinaryOp::BitAnd => match (src, dst) {
+                (Operand::Reg(s), Operand::Reg(d)) => a.and(gpr32(d), gpr32(s))?,
+                (Operand::Imm(v), Operand::Reg(d)) => a.and(gpr32(d), *v as i32)?,
+                (Operand::Reg(s), Operand::Stack(off)) => a.and(mem_rbp(*off), gpr32(s))?,
+                (Operand::Imm(v), Operand::Stack(off)) => a.and(mem_rbp(*off), *v as i32)?,
+                _ => unimplemented!(),
+            },
+            BinaryOp::BitOr => match (src, dst) {
+                (Operand::Reg(s), Operand::Reg(d)) => a.or(gpr32(d), gpr32(s))?,
+                (Operand::Imm(v), Operand::Reg(d)) => a.or(gpr32(d), *v as i32)?,
+                (Operand::Reg(s), Operand::Stack(off)) => a.or(mem_rbp(*off), gpr32(s))?,
+                (Operand::Imm(v), Operand::Stack(off)) => a.or(mem_rbp(*off), *v as i32)?,
+                _ => unimplemented!(),
+            },
+            BinaryOp::BitXOr => match (src, dst) {
+                (Operand::Reg(s), Operand::Reg(d)) => a.xor(gpr32(d), gpr32(s))?,
+                (Operand::Imm(v), Operand::Reg(d)) => a.xor(gpr32(d), *v as i32)?,
+                (Operand::Reg(s), Operand::Stack(off)) => a.xor(mem_rbp(*off), gpr32(s))?,
+                (Operand::Imm(v), Operand::Stack(off)) => a.xor(mem_rbp(*off), *v as i32)?,
+                _ => unimplemented!(),
+            },
+            BinaryOp::BitShl => match (src, dst) {
+                (Operand::Imm(v), Operand::Reg(d)) => a.shl(gpr32(d), *v as i32)?,
+                (Operand::Reg(_s), Operand::Reg(d)) => a.shl(gpr32(d), gpr8::cl)?,
+                (Operand::Imm(v), Operand::Stack(off)) => a.shl(mem_rbp(*off), *v as i32)?,
+                (Operand::Reg(s), Operand::Stack(off)) => {
+                    a.mov(registers::gpr32::ecx, gpr32(s))?;
+                    a.shl(mem_rbp(*off), gpr8::cl)?;
+                }
+                _ => unimplemented!(),
+            },
+            BinaryOp::BitSar => match (src, dst) {
+                (Operand::Imm(v), Operand::Reg(d)) => a.sar(gpr32(d), *v as i32)?,
+                (Operand::Reg(_s), Operand::Reg(d)) => a.sar(gpr32(d), gpr8::cl)?,
+                (Operand::Imm(v), Operand::Stack(off)) => a.sar(mem_rbp(*off), *v as i32)?,
+                (Operand::Reg(s), Operand::Stack(off)) => {
+                    a.mov(registers::gpr32::ecx, gpr32(s))?;
+                    a.sar(mem_rbp(*off), gpr8::cl)?;
+                }
+                _ => unimplemented!(),
+            },
+        },
+        Instruction::Cmp { v1, v2 } => match (v1, v2) {
+            (Operand::Reg(r1), Operand::Reg(r2)) => a.cmp(gpr32(r2), gpr32(r1))?,
+            (Operand::Reg(r1), Operand::Imm(v)) => a.cmp(gpr32(r1), *v as i32)?,
+            (Operand::Reg(r1), Operand::Stack(off)) => a.cmp(mem_rbp(*off), gpr32(r1))?,
+            (Operand::Stack(off), Operand::Reg(r)) => a.cmp(gpr32(r), mem_rbp(*off))?,
+            (Operand::Stack(o1), Operand::Imm(v)) => a.cmp(mem_rbp(*o1), *v as i32)?,
+            (Operand::Imm(v), Operand::Reg(r)) => a.cmp(gpr32(r), *v as i32)?,
+            (Operand::Imm(v), Operand::Stack(off)) => a.cmp(mem_rbp(*off), *v as i32)?,
+            _ => unimplemented!(),
+        },
+        Instruction::Cdq => {
+            a.cdq()?;
+        }
+        Instruction::Idiv(op) => match op {
+            Operand::Reg(r) => a.idiv(gpr32(r))?,
+            Operand::Stack(off) => a.idiv(mem_rbp(*off))?,
+            _ => unimplemented!(),
+        },
+        Instruction::Jmp(label) => {
+            let l = *labels.entry(label.0.clone()).or_insert_with(|| a.create_label());
+            a.jmp(l)?;
+        }
+        Instruction::JmpCC { code, label } => {
+            let l = *labels.entry(label.0.clone()).or_insert_with(|| a.create_label());
+            match code {
+                CondCode::E => a.je(l)?,
+                CondCode::NE => a.jne(l)?,
+                CondCode::G => a.jg(l)?,
+                CondCode::GE => a.jge(l)?,
+                CondCode::L => a.jl(l)?,
+                CondCode::LE => a.jle(l)?,
+            };
+        }
+        Instruction::SetCC { code, op } => match op {
+            Operand::Reg(_r) => match code {
+                CondCode::E => a.sete(gpr8::al)?,
+                CondCode::NE => a.setne(gpr8::al)?,
+                CondCode::G => a.setg(gpr8::al)?,
+                CondCode::GE => a.setge(gpr8::al)?,
+                CondCode::L => a.setl(gpr8::al)?,
+                CondCode::LE => a.setle(gpr8::al)?,
+            },
+            Operand::Stack(off) => match code {
+                CondCode::E => a.sete(mem_rbp(*off))?,
+                CondCode::NE => a.setne(mem_rbp(*off))?,
+                CondCode::G => a.setg(mem_rbp(*off))?,
+                CondCode::GE => a.setge(mem_rbp(*off))?,
+                CondCode::L => a.setl(mem_rbp(*off))?,
+                CondCode::LE => a.setle(mem_rbp(*off))?,
+            },
+            _ => unimplemented!(),
+        },
+        Instruction::Label(lbl) => {
+            let l = labels.entry(lbl.0.clone()).or_insert_with(|| a.create_label());
+            a.set_label(l)?;
+        }
+        Instruction::AllocateStack(off) => {
+            if *off != 0 {
+                a.sub(gpr64::rsp, -*off as i32)?;
+            }
+        }
+        Instruction::Ret => {
+            a.mov(gpr64::rsp, gpr64::rbp)?;
+            a.pop(gpr64::rbp)?;
+            a.ret()?;
+        }
+    }
+    Ok(())
 }
