+++ src/main.rs
@@ -1,6 +1,5 @@
 mod codegen;
 mod emit;
-mod emit_iced;
 mod lexer;
 mod parser;
 mod pretty;
@@ -8,7 +7,8 @@ mod tacky;
 
 use crate::pretty::ItfDisplay;
 use clap::{ArgGroup, Parser};
-use iced_x86::Formatter;
+use colored::{ColoredString, Colorize};
+use iced_x86::{Formatter, FormatterOutput, FormatterTextKind};
 use libwild::{Args as WildArgs, Linker};
 use std::fs;
 use std::path::Path;
@@ -45,7 +45,7 @@ struct Args {
     #[arg(long, name = "run")]
     run: bool,
 
-    /// Use GCC for assembling and linking
+    /// Use GCC for linking
     #[arg(long)]
     gcc: bool,
 
@@ -56,6 +56,35 @@ struct Args {
     filename: String,
 }
 
+// For colored output, we need to implement a custom FormatterOutput
+struct MyFormatterOutput {
+    vec: Vec<(String, FormatterTextKind)>,
+}
+impl MyFormatterOutput {
+    pub fn new() -> Self {
+        Self { vec: Vec::new() }
+    }
+}
+
+impl FormatterOutput for MyFormatterOutput {
+    fn write(&mut self, text: &str, kind: FormatterTextKind) {
+        // This allocates a string. If that's a problem, just call print!() here
+        // instead of storing the result in a vector.
+        self.vec.push((String::from(text), kind));
+    }
+}
+
+fn get_color(s: &str, kind: FormatterTextKind) -> ColoredString {
+    match kind {
+        FormatterTextKind::Directive | FormatterTextKind::Keyword => s.bright_yellow(),
+        FormatterTextKind::Prefix | FormatterTextKind::Mnemonic => s.bright_magenta(),
+        FormatterTextKind::Register => s.bright_blue(),
+        FormatterTextKind::Number => s.bright_cyan(),
+        _ => s.white(),
+    }
+}
+
+// todo default to external linker on macOS
 fn main() {
     let args = Args::parse();
     let input = fs::read_to_string(&args.filename).expect("Failed to read input file");
@@ -104,44 +133,37 @@ fn main() {
         .output
         .unwrap_or_else(|| path.with_extension("").to_string_lossy().to_string());
 
-    if !args.gcc {
-        // use iced_x86 for assembly generation
-        let obj = emit_iced::emit_object(&code_ast).expect("iced obj");
+    let obj = emit::emit_object(&code_ast).expect("iced obj");
 
-        if args.s {
-            let mut formatter = iced_x86::GasFormatter::new();
-            let mut output = String::new();
+    if args.s {
+        let mut formatter = iced_x86::GasFormatter::new();
+        let ops = formatter.options_mut();
+        ops.set_number_base(iced_x86::NumberBase::Decimal);
+        let mut output = MyFormatterOutput::new();
 
-            println!("Generated ASM");
-            let asm = emit_iced::get_instructions(&code_ast).unwrap();
-            for ins in asm {
-                output.clear();
-                formatter.format(&ins, &mut output);
-                println!("{output}");
+        println!("Generated ASM");
+        let asm = emit::get_instructions(&code_ast).unwrap();
+        for ins in asm {
+            output.vec.clear();
+            formatter.format(&ins, &mut output);
+            for (text, kind) in output.vec.iter() {
+                print!("{}", get_color(text.as_str(), *kind));
             }
+            println!();
         }
+    }
 
-        let obj_file = format!("{out_file}.o");
-        fs::write(&obj_file, &obj).expect("Failed to write object file");
-        let linker = Linker::new();
-        let args_vec = ["-o", &out_file, &obj_file, "--entry=_start"];
-        let parsed = WildArgs::parse(args_vec.iter()).expect("parse args");
-        let _ = libwild::setup_tracing(&parsed);
-        linker.run(&parsed).expect("link failed");
-        fs::remove_file(&obj_file).ok();
-    } else {
-        let asm = emit::emit_program(&code_ast);
-        if args.s {
-            print!("Generated asm:\n\n{asm}");
-        }
-        let asm_file = format!("{out_file}.s");
+    let obj_file = format!("{out_file}.o");
+    fs::write(&obj_file, &obj).expect("Failed to write object file");
 
-        fs::write(&asm_file, asm).expect("Failed to write assembly file");
-
+    if args.gcc {
         let status = std::process::Command::new("gcc")
-            .arg(&asm_file)
+            .arg("-nostdlib")
+            .arg("-static")
+            .arg(&obj_file)
             .arg("-o")
-            .arg(out_file.clone())
+            .arg(&out_file)
+            .arg("-Wl,-e,_start")
             .status()
             .expect("Failed to execute gcc");
 
@@ -149,14 +171,12 @@ fn main() {
             println!("Compilation failed with status: {status}");
             std::process::exit(1);
         }
-
-        fs::remove_file(&asm_file).expect("Failed to delete assembly file");
+    } else {
+        let linker = Linker::new();
+        let args_vec = ["-o", &out_file, &obj_file, "--entry=_start"];
+        let parsed = WildArgs::parse(args_vec.iter()).expect("parse args");
+        let _ = libwild::setup_tracing(&parsed);
+        linker.run(&parsed).expect("link failed");
     }
-
-    if args.run {
-        let run_status = std::process::Command::new(&out_file)
-            .status()
-            .expect("Failed to execute compiled binary");
-        println!("Result: {}", run_status.code().unwrap());
-    }
+    fs::remove_file(&obj_file).ok();
 }
